Renamed and moved travis scripts to travis folder
