+++ .travis.yml
@@ -8,9 +8,10 @@ jobs:
     - if: type = push AND NOT branch = master
       script:
         - echo "Ordinary push"
-        - python ./scripts/travis_env_test.py
+        - python ./scripts/travis/env_test.py
+        
     - if: type = push AND branch = master
       install:
-        - pip install -r requirements.txt
+        - pip install -r ./scripts/travis/requirements.txt
       script:
-        - python ./scripts/travis.py
+        - python ./scripts/travis/deploy.py
