--- scripts/travis_env_test.py
@@ -1,4 +1,0 @@
-from travis import createCertificate
-
-if __name__=='__main__':
-    print(createCertificate())
